Build an app that recognizes faces. Let's begin the workshop with this lab.
The Image Analysis service is trained on over 1 million faces, rather famous people. This service reveals the age range and the gender of any uploaded face (or faces), and if it happens to be one that the system is trained on, it will identify the individual or group of people, with the confidence measure next to each parameter.